Renames SEO and About files so imports are cleaner.
