+++ src/pages/index.tsx
@@ -1,9 +1,10 @@
 import React from 'react';
-import Layout from '../App/Layout/Layout';
-import SEO from '../App/SEO/SEO';
-import Header from '../App/Header';
 
-import About from '../App/About/About';
+// Components
+import About from '../App/About';
+import Layout from '../App/Layout/Layout';
+import Header from '../App/Header';
+import SEO from '../App/SEO';
 
 const IndexPage: React.FC = () => {
   return (
